split off sensor code into class
