+++ SensorClient/src/main.cpp
@@ -14,19 +14,20 @@
 #include <Wire.h> // I2C library
 #include <ezTime.h>
 
+#include "common_sensor.h"
 #include "config.h"
 #include "logging.h"
 
 #ifdef HAS_AM2320
-#include <AM232X.h>
+#include <AM2320Sensor.h>
 #endif
 
 #ifdef HAS_CCS811
-#include "ccs811.h" // CCS811 library
+#include "CCS811Sensor.h"
 #endif
 
 #ifdef HAS_HDC1080
-#include "ClosedCube_HDC1080.h" // HDC1080 library
+#include "ClosedCube_HDC1080.h"
 #endif
 
 #ifdef HAS_HP303B
@@ -96,576 +97,6 @@ uint8 num_sending_measurement_errors = 0;
 
 //// Time
 Timezone Amsterdam;
-
-///// Common sensor
-typedef struct {
-  time_t epoch;
-  uint8_t sensor_id;
-  uint8_t magnitude_id;
-  char value[10];
-} SensorData;
-SensorData sensor_data;
-CircularBuffer<SensorData, 255> sensor_buffer; // Keep some raw data
-uint8 num_measurement_errors = 0;
-
-#ifdef HAS_AM2320
-// AM2320 Sensor
-uint8_t am2320_sensor_id, am2320_temp_id, am2320_hum_id;
-const char *sensor_am2320_name PROGMEM = "AM2320";
-AM232X am2320;
-const uint32_t am2320_period_s = 10;
-const size_t capacity_am2320 =
-    JSON_ARRAY_SIZE(2) + JSON_OBJECT_SIZE(2) + 2 * JSON_OBJECT_SIZE(3);
-const size_t response_capacity_am2320 =
-    JSON_ARRAY_SIZE(2) + JSON_OBJECT_SIZE(3) + 2 * JSON_OBJECT_SIZE(4);
-
-void setup_am2320_sensor_json(JsonObject &sensor_json) {
-  sensor_json["name"] = sensor_am2320_name;
-  JsonArray sensor_1_in_magnitudes =
-      sensor_json.createNestedArray("magnitudes");
-
-  JsonObject mag1_in = sensor_1_in_magnitudes.createNestedObject();
-  mag1_in["name"] = "temperature";
-  mag1_in["unit"] = "C";
-  mag1_in["precision"] = 0.1;
-
-  JsonObject mag2_in = sensor_1_in_magnitudes.createNestedObject();
-  mag2_in["name"] = "humidity";
-  mag2_in["unit"] = "%";
-  mag2_in["precision"] = 0.1;
-}
-
-bool parse_am2320_sensor_json(JsonObject &sensor_json_response) {
-  am2320_sensor_id = sensor_json_response["id"];
-  JsonArray magnitudes_json =
-      sensor_json_response["magnitudes"].as<JsonArray>();
-  for (JsonObject mag_json : magnitudes_json) {
-    const char *name = mag_json["name"];
-    if (strcmp(name, "temperature") == 0) {
-      am2320_temp_id = mag_json["id"];
-    } else if (strcmp(name, "humidity") == 0) {
-      am2320_hum_id = mag_json["id"];
-    }
-  }
-  log_printf("  am2320_sensor_id: %d, am2320_temp_id: %d, am2320_hum_id: %d.\n",
-             am2320_sensor_id, am2320_temp_id, am2320_hum_id);
-  log_header_printf(
-      "  AM2320 sensor_id: %d, temperature_id: %d, humidity_id: %d.",
-      am2320_sensor_id, am2320_temp_id, am2320_hum_id);
-
-  return true;
-}
-
-#ifdef HAS_CCS811
-void correct_ccs811_envdata(float humidity, float temperature);
-#endif
-
-bool setup_am2320_sensor() {
-  log_println("Setting up AM2320 sensor...");
-  if (!am2320.begin()) {
-    delay(100);
-    if (!am2320.isConnected()) {
-      log_println("  AM2320 begin FAILED");
-      return false;
-    }
-  }
-  delay(500);
-
-  log_printf("  Model: %d.\n"
-             "  Version: %d.\n"
-             "  Device ID: %d.\n"
-             "  Done!\n",
-             am2320.getModel(), am2320.getVersion(), am2320.getDeviceID());
-  log_header_printf("AM2320 setup. model: %d, "
-                    "version: %d, dev. id: %d.",
-                    am2320.getModel(), am2320.getVersion(),
-                    am2320.getDeviceID());
-
-  return true;
-}
-
-void measure_am2320_sensor() {
-  log_println("Measuring AM2320...");
-  int conversion_ret_val;
-  time_t now = UTC.now();
-  float temperature, humidity;
-
-  int status = am2320.read();
-  if (status != AM232X_OK) {
-    log_printf("  Error reading AM2320 (%d).\n", status);
-    num_measurement_errors++;
-    return;
-  }
-
-  // sensor
-  humidity = am2320.getHumidity();
-  temperature = am2320.getTemperature();
-
-  SensorData humidity_data = {now, am2320_sensor_id, am2320_hum_id};
-  conversion_ret_val = snprintf(humidity_data.value,
-                                sizeof(humidity_data.value), "%.3f", humidity);
-  if (conversion_ret_val > 0) {
-    log_printf("  Humidity: %.2f %%.\n", humidity);
-    sensor_buffer.push(humidity_data);
-    if (num_measurement_errors > 0) {
-      num_measurement_errors--;
-    }
-  } else {
-    log_printf("  Problem converting humidity (%.2f %%) to char[].\n",
-               humidity);
-  }
-
-  SensorData temperature_data = {now, am2320_sensor_id, am2320_temp_id};
-  conversion_ret_val =
-      snprintf(temperature_data.value, sizeof(temperature_data.value), "%.3f",
-               temperature);
-  if (conversion_ret_val > 0) {
-    log_printf("  Temperature: %.2f C.\n", temperature);
-    sensor_buffer.push(temperature_data);
-    if (num_measurement_errors > 0) {
-      num_measurement_errors--;
-    }
-  } else {
-    log_printf("  Problem converting temperature (%.2f C) to char[].\n",
-               temperature);
-  }
-
-#ifdef HAS_CCS811
-  correct_ccs811_envdata(humidity, temperature);
-#endif
-}
-
-Ticker am2320_measurement_timer(measure_am2320_sensor, am2320_period_s * 1e3, 0,
-                                MILLIS);
-#endif
-
-#ifdef HAS_CCS811
-// CCS811 Sensor
-uint8_t ccs811_sensor_id, ccs811_eco2_id, ccs811_etvoc_id;
-const char *sensor_ccs811_name PROGMEM = "CCS811";
-CCS811 ccs811;
-const uint32_t ccs811_period_s = 10;
-const size_t capacity_ccs811 =
-    JSON_ARRAY_SIZE(2) + JSON_OBJECT_SIZE(2) + 2 * JSON_OBJECT_SIZE(3);
-const size_t response_capacity_ccs811 =
-    JSON_ARRAY_SIZE(2) + JSON_OBJECT_SIZE(3) + 2 * JSON_OBJECT_SIZE(4);
-
-void setup_ccs811_sensor_json(JsonObject &sensor_json) {
-  sensor_json["name"] = sensor_ccs811_name;
-  JsonArray sensor_1_in_magnitudes =
-      sensor_json.createNestedArray("magnitudes");
-
-  JsonObject mag1_in = sensor_1_in_magnitudes.createNestedObject();
-  mag1_in["name"] = "eco2";
-  mag1_in["unit"] = "ppm";
-  mag1_in["precision"] = 1;
-
-  JsonObject mag2_in = sensor_1_in_magnitudes.createNestedObject();
-  mag2_in["name"] = "etvoc";
-  mag2_in["unit"] = "ppb";
-  mag2_in["precision"] = 1;
-}
-
-bool parse_ccs811_sensor_json(JsonObject &sensor_json_response) {
-  ccs811_sensor_id = sensor_json_response["id"];
-  JsonArray magnitudes_json =
-      sensor_json_response["magnitudes"].as<JsonArray>();
-  for (JsonObject mag_json : magnitudes_json) {
-    const char *name = mag_json["name"];
-    if (strcmp(name, "eco2") == 0) {
-      ccs811_eco2_id = mag_json["id"];
-    } else if (strcmp(name, "etvoc") == 0) {
-      ccs811_etvoc_id = mag_json["id"];
-    }
-  }
-  log_printf(
-      "  ccs811_sensor_id: %d, ccs811_eco2_id: %d, ccs811_etvoc_id: %d.\n",
-      ccs811_sensor_id, ccs811_eco2_id, ccs811_etvoc_id);
-  log_header_printf("  CCS811 sensor_id: %d, eCO2_id: %d, eTVOC_id: %d.",
-                    ccs811_sensor_id, ccs811_eco2_id, ccs811_etvoc_id);
-
-  return true;
-}
-
-bool setup_ccs811_sensor() {
-  log_printf("Setting up CCS811 sensor (version %d)...\n", CCS811_VERSION);
-
-  // Enable CCS811
-  ccs811.set_i2cdelay(20); // Needed for ESP8266 because it doesn't handle I2C
-                           // clock stretch correctly
-  bool ok = ccs811.begin();
-  if (!ok) {
-    log_println("  CCS811 begin FAILED");
-    return false;
-  }
-
-  // Print CCS811 versions
-  log_printf("  hardware version: %X\n", ccs811.hardware_version());
-  log_printf("  bootloader version: %X\n", ccs811.bootloader_version());
-  log_printf("  application version: %X\n", ccs811.application_version());
-
-  // Start measuring
-  ok = ccs811.start(CCS811_MODE_10SEC);
-  if (!ok) {
-    log_println("  CCS811 start FAILED");
-    return false;
-  }
-
-  delay(500);
-  log_println("  Done!");
-  log_header_printf("CCS811 setup. lib v. %d, hw v.: 0x%X, "
-                    "bootldr v.: 0x%X, app v.: 0x%X.",
-                    CCS811_VERSION, ccs811.hardware_version(),
-                    ccs811.bootloader_version(), ccs811.application_version());
-  return true;
-}
-
-void measure_ccs811_sensor() {
-  log_println("Measuring CCS811...");
-  int conversion_ret_val;
-  time_t now = UTC.now();
-
-  // sensor
-  uint16_t eco2, etvoc, errstat, raw;
-  ccs811.read(&eco2, &etvoc, &errstat, &raw);
-
-  // Check for errors
-  if (errstat != CCS811_ERRSTAT_OK) {
-    num_measurement_errors++;
-    if (errstat == CCS811_ERRSTAT_OK_NODATA) {
-      log_println("  error: waiting for (new) data");
-    } else if (errstat & CCS811_ERRSTAT_I2CFAIL) {
-      log_println("  I2C error");
-    } else {
-      log_printf("  other error: errstat (%X) =", errstat);
-      log_println(ccs811.errstat_str(errstat));
-    }
-    return;
-  }
-
-  SensorData eco2_data = {now, ccs811_sensor_id, ccs811_eco2_id};
-  conversion_ret_val =
-      snprintf(eco2_data.value, sizeof(eco2_data.value), "%d", eco2);
-  if (conversion_ret_val > 0) {
-    log_printf("  equivalent CO2: %d ppm.\n", eco2);
-    sensor_buffer.push(eco2_data);
-    if (num_measurement_errors > 0) {
-      num_measurement_errors--;
-    }
-  } else {
-    log_printf("  Problem converting equivalent CO2 (%d ppm) to char[].\n",
-               eco2);
-  }
-
-  SensorData etvoc_data = {now, ccs811_sensor_id, ccs811_etvoc_id};
-  conversion_ret_val =
-      snprintf(etvoc_data.value, sizeof(etvoc_data.value), "%d", etvoc);
-  if (conversion_ret_val > 0) {
-    log_printf("  total VOC: %d ppb.\n", etvoc);
-    sensor_buffer.push(etvoc_data);
-    if (num_measurement_errors > 0) {
-      num_measurement_errors--;
-    }
-  } else {
-    log_printf("  Problem converting total VOC (%d ppb) to char[].\n", etvoc);
-  }
-}
-
-void correct_ccs811_envdata(float humidity, float temperature) {
-#ifdef HAS_CCS811
-  ccs811.set_envdata(((uint16_t)humidity + 250) / 500,
-                     ((uint16_t)temperature + 25250) / 500);
-#endif
-}
-Ticker ccs811_measurement_timer(measure_ccs811_sensor, ccs811_period_s * 1e3, 0,
-                                MILLIS);
-#endif
-
-#ifdef HAS_HDC1080
-// HAS_HDC1080 Sensor
-uint8_t hdc1080_sensor_id, hdc1080_temp_id, hdc1080_hum_id;
-const char *sensor_hdc1080_name PROGMEM = "HDC1080";
-ClosedCube_HDC1080 hdc1080;
-const uint32_t hdc1080_period_s = 10;
-const size_t capacity_hdc1080 =
-    JSON_ARRAY_SIZE(2) + JSON_OBJECT_SIZE(2) + 2 * JSON_OBJECT_SIZE(3);
-const size_t response_capacity_hdc1080 =
-    JSON_ARRAY_SIZE(2) + JSON_OBJECT_SIZE(3) + 2 * JSON_OBJECT_SIZE(4);
-
-void setup_hdc1080_sensor_json(JsonObject &sensor_json) {
-  sensor_json["name"] = sensor_hdc1080_name;
-  JsonArray sensor_1_in_magnitudes =
-      sensor_json.createNestedArray("magnitudes");
-
-  JsonObject mag1_in = sensor_1_in_magnitudes.createNestedObject();
-  mag1_in["name"] = "temperature";
-  mag1_in["unit"] = "C";
-  mag1_in["precision"] = 0.2;
-
-  JsonObject mag2_in = sensor_1_in_magnitudes.createNestedObject();
-  mag2_in["name"] = "humidity";
-  mag2_in["unit"] = "%";
-  mag2_in["precision"] = 2;
-}
-
-bool parse_hdc1080_sensor_json(JsonObject &sensor_json_response) {
-  hdc1080_sensor_id = sensor_json_response["id"];
-  JsonArray magnitudes_json =
-      sensor_json_response["magnitudes"].as<JsonArray>();
-  for (JsonObject mag_json : magnitudes_json) {
-    const char *name = mag_json["name"];
-    if (strcmp(name, "temperature") == 0) {
-      hdc1080_temp_id = mag_json["id"];
-    } else if (strcmp(name, "humidity") == 0) {
-      hdc1080_hum_id = mag_json["id"];
-    }
-  }
-  log_printf(
-      "  hdc1080_sensor_id: %d, hdc1080_temp_id: %d, hdc1080_hum_id: %d.\n",
-      hdc1080_sensor_id, hdc1080_temp_id, hdc1080_hum_id);
-  log_header_printf(
-      "  HDC1080 sensor_id: %d, temperature_id: %d, humidity_id: %d.",
-      hdc1080_sensor_id, hdc1080_temp_id, hdc1080_hum_id);
-
-  return true;
-}
-
-bool setup_hdc1080_sensor() {
-  log_printf("Setting up HDC1080 sensor...\n");
-
-  // Enable HDC1080
-  hdc1080.begin(0x40);
-  hdc1080.setResolution(HDC1080_RESOLUTION_14BIT, HDC1080_RESOLUTION_14BIT);
-
-  if (hdc1080.readManufacturerId() == 0xFFFF) {
-    log_println("  Communication error!");
-    return false;
-  }
-
-  // Print HDC1080 versions
-  log_printf("  Manufacturer ID: 0x%X\n", hdc1080.readManufacturerId());
-  log_printf("  Device ID: 0x%X\n", hdc1080.readDeviceId());
-  HDC1080_SerialNumber sernum = hdc1080.readSerialNumber();
-  log_printf("  Serial number: %02X-%04X-%04X.\n", sernum.serialFirst,
-             sernum.serialMid, sernum.serialLast);
-
-  delay(500);
-  log_println("  Done!");
-  log_header_printf("HDC1080 manufacturer ID v. 0x%X, dev ID: 0x%X, "
-                    "serial no.: %02X-%04X-%04X.",
-                    hdc1080.readManufacturerId(), hdc1080.readDeviceId(),
-                    sernum.serialFirst, sernum.serialMid, sernum.serialLast);
-  return true;
-}
-
-void measure_hdc1080_sensor() {
-  log_println("Measuring HDC1080...");
-  int conversion_ret_val;
-  time_t now = UTC.now();
-  float temperature, humidity;
-
-  // sensor
-  humidity = hdc1080.readHumidity();
-  temperature = hdc1080.readTemperature();
-
-  if (isnan(humidity)) {
-    log_println("  Error reading humidity.");
-    num_measurement_errors++;
-  } else if (humidity > 99.99) {
-    log_printf("  Error reading humidity (%.1f).\n", humidity);
-    num_measurement_errors++;
-  } else {
-    SensorData humidity_data = {now, hdc1080_sensor_id, hdc1080_hum_id};
-    conversion_ret_val = snprintf(
-        humidity_data.value, sizeof(humidity_data.value), "%.1f", humidity);
-    if (conversion_ret_val > 0) {
-      log_printf("  Humidity: %.1f %%.\n", humidity);
-      sensor_buffer.push(humidity_data);
-      if (num_measurement_errors > 0) {
-        num_measurement_errors--;
-      }
-    } else {
-      log_printf("  Problem converting humidity (%.1f %%) to char[].\n",
-                 humidity);
-    }
-  }
-
-  if (isnan(temperature)) {
-    log_println("  Error reading temperature.");
-    num_measurement_errors++;
-  } else if (temperature > 120) {
-    log_printf("  Error reading temperature (%.2f).\n", temperature);
-    num_measurement_errors++;
-  } else {
-    SensorData temperature_data = {now, hdc1080_sensor_id, hdc1080_temp_id};
-    conversion_ret_val =
-        snprintf(temperature_data.value, sizeof(temperature_data.value), "%.2f",
-                 temperature);
-    if (conversion_ret_val > 0) {
-      log_printf("  Temperature: %.2f C.\n", temperature);
-      sensor_buffer.push(temperature_data);
-      if (num_measurement_errors > 0) {
-        num_measurement_errors--;
-      }
-    } else {
-      log_printf("  Problem converting temperature (%.2f C) to char[].\n",
-                 temperature);
-    }
-  }
-}
-
-Ticker hdc1080_measurement_timer(measure_hdc1080_sensor, hdc1080_period_s * 1e3,
-                                 0, MILLIS);
-#endif
-
-#ifdef HAS_HP303B
-// HP303B Sensor
-uint8_t hp303b_sensor_id, hp303b_temp_id, hp303b_pres_id;
-const char *sensor_hp303b_name PROGMEM = "HP303B";
-LOLIN_HP303B hp303b;
-const uint32_t hp303b_period_s = 10;
-const size_t capacity_hp303b =
-    JSON_ARRAY_SIZE(2) + JSON_OBJECT_SIZE(2) + 2 * JSON_OBJECT_SIZE(3);
-const size_t response_capacity_hp303b =
-    JSON_ARRAY_SIZE(2) + JSON_OBJECT_SIZE(3) + 2 * JSON_OBJECT_SIZE(4);
-
-void setup_hp303b_sensor_json(JsonObject &sensor_json) {
-  sensor_json["name"] = sensor_hp303b_name;
-  JsonArray sensor_1_in_magnitudes =
-      sensor_json.createNestedArray("magnitudes");
-
-  JsonObject mag1_in = sensor_1_in_magnitudes.createNestedObject();
-  mag1_in["name"] = "temperature";
-  mag1_in["unit"] = "C";
-  mag1_in["precision"] = 0.5;
-
-  JsonObject mag2_in = sensor_1_in_magnitudes.createNestedObject();
-  mag2_in["name"] = "pressure";
-  mag2_in["unit"] = "Pa";
-  mag2_in["precision"] = 10;
-}
-
-bool parse_hp303b_sensor_json(JsonObject &sensor_json_response) {
-  hp303b_sensor_id = sensor_json_response["id"];
-  JsonArray magnitudes_json =
-      sensor_json_response["magnitudes"].as<JsonArray>();
-  for (JsonObject mag_json : magnitudes_json) {
-    const char *name = mag_json["name"];
-    if (strcmp(name, "temperature") == 0) {
-      hp303b_temp_id = mag_json["id"];
-    } else if (strcmp(name, "pressure") == 0) {
-      hp303b_pres_id = mag_json["id"];
-    }
-  }
-  log_printf(
-      "  hp303b_sensor_id: %d, hp303b_temp_id: %d, hp303b_pres_id: %d.\n",
-      hp303b_sensor_id, hp303b_temp_id, hp303b_pres_id);
-  log_header_printf(
-      "  HP303B sensor_id: %d, temperature_id: %d, pressure_id: %d.",
-      hp303b_sensor_id, hp303b_temp_id, hp303b_pres_id);
-
-  return true;
-}
-
-bool setup_hp303b_sensor() {
-  log_println("Setting up HP303B sensor...");
-  hp303b.begin();
-  delay(500);
-
-  log_printf("  Product ID: %d.\n"
-             "  Revision ID: %d.\n"
-             "  Done!\n",
-             hp303b.getProductId(), hp303b.getRevisionId());
-  log_header_printf("HP303B setup. product ID: %d, "
-                    "revision ID: %d.",
-                    hp303b.getProductId(), hp303b.getRevisionId());
-
-  return true;
-}
-
-void measure_hp303b_sensor() {
-  log_println("Measuring HP303B...");
-  const uint8_t oversampling = 7;
-  int conversion_ret_val;
-  time_t now = UTC.now();
-  int temperature, pressure;
-  int status;
-
-  status = hp303b.measureTempOnce(temperature, oversampling);
-  switch (status) {
-  case HP303B__SUCCEEDED:
-    break;
-  case HP303B__FAIL_UNKNOWN:
-    log_println("  Unknown error reading HP303B.");
-    num_measurement_errors++;
-    break;
-  case HP303B__FAIL_INIT_FAILED:
-    log_println("  Initialization error reading HP303B.");
-    num_measurement_errors++;
-    break;
-  case HP303B__FAIL_TOOBUSY:
-    log_println("  Error: HP303B is busy.");
-    num_measurement_errors++;
-    break;
-  case HP303B__FAIL_UNFINISHED:
-    log_println("  Error: HP303B could not finish the measurement on time.");
-    num_measurement_errors++;
-    break;
-  }
-  SensorData temperature_data = {now, hp303b_sensor_id, hp303b_temp_id};
-  conversion_ret_val =
-      snprintf(temperature_data.value, sizeof(temperature_data.value), "%d",
-               temperature);
-  if (conversion_ret_val > 0) {
-    log_printf("  Temperature: %d C.\n", temperature);
-    sensor_buffer.push(temperature_data);
-    if (num_measurement_errors > 0) {
-      num_measurement_errors--;
-    }
-  } else {
-    log_printf("  Problem converting temperature (%d C) to char[].\n",
-               temperature);
-  }
-
-  status = hp303b.measurePressureOnce(pressure, oversampling);
-  switch (status) {
-  case HP303B__SUCCEEDED:
-    break;
-  case HP303B__FAIL_UNKNOWN:
-    log_println("  Unknown error reading HP303B.");
-    num_measurement_errors++;
-    break;
-  case HP303B__FAIL_INIT_FAILED:
-    log_println("  Initialization error reading HP303B.");
-    num_measurement_errors++;
-    break;
-  case HP303B__FAIL_TOOBUSY:
-    log_println("  Error: HP303B is busy.");
-    num_measurement_errors++;
-    break;
-  case HP303B__FAIL_UNFINISHED:
-    log_println("  Error: HP303B could not finish the measurement on time.");
-    num_measurement_errors++;
-    break;
-  }
-  SensorData pressure_data = {now, hp303b_sensor_id, hp303b_pres_id};
-  conversion_ret_val = snprintf(pressure_data.value,
-                                sizeof(pressure_data.value), "%d", pressure);
-  if (conversion_ret_val > 0) {
-    log_printf("  Pressure: %d hPa.\n", pressure / 100);
-    sensor_buffer.push(pressure_data);
-    if (num_measurement_errors > 0) {
-      num_measurement_errors--;
-    }
-  } else {
-    log_printf("  Problem converting pressure (%d hPa) to char[].\n",
-               pressure / 100);
-  }
-}
-
-Ticker hp303b_measurement_timer(measure_hp303b_sensor, hp303b_period_s * 1e3, 0,
-                                MILLIS);
-#endif
 
 //// Post request
 WiFiClient client;
@@ -835,7 +266,7 @@ bool setup_sensors() {
   // Prepare JSON document
   size_t capacity = JSON_ARRAY_SIZE(NUM_SENSORS);
 #ifdef HAS_AM2320
-  capacity += capacity_am2320;
+  capacity += am2320_sensor.capacity;
 #endif
 #ifdef HAS_CCS811
   capacity += capacity_ccs811;
@@ -850,7 +281,7 @@ bool setup_sensors() {
 
   size_t response_capacity = JSON_ARRAY_SIZE(NUM_SENSORS);
 #ifdef HAS_AM2320
-  response_capacity += response_capacity_am2320;
+  response_capacity += am2320_sensor.response_capacity;
 #endif
 #ifdef HAS_CCS811
   response_capacity += response_capacity_ccs811;
@@ -865,7 +296,7 @@ bool setup_sensors() {
 
 #ifdef HAS_AM2320
   JsonObject am2320_json = sensors_json.createNestedObject();
-  setup_am2320_sensor_json(am2320_json);
+  am2320_sensor.setup_json(am2320_json);
 #endif
 #ifdef HAS_CCS811
   JsonObject ccs811_json = sensors_json.createNestedObject();
@@ -913,8 +344,8 @@ bool setup_sensors() {
   for (JsonObject sensor_json : sensors_json_out) {
     const char *name = sensor_json["name"];
 #ifdef HAS_AM2320
-    if (strcmp(name, sensor_am2320_name) == 0) {
-      parse_am2320_sensor_json(sensor_json);
+    if (strcmp(name, am2320_sensor.name) == 0) {
+      am2320_sensor.parse_json(sensor_json);
     }
 #endif
 #ifdef HAS_CCS811
@@ -940,7 +371,7 @@ bool setup_sensors() {
 bool setup_internal_sensors() {
   bool res = true;
 #ifdef HAS_AM2320
-  res = res && setup_am2320_sensor();
+  res = res && am2320_sensor.setup();
 #endif
 
 #ifdef HAS_HDC1080
